docs: axios_config_and_sweetalert
