+++ package.json
@@ -26,10 +26,12 @@
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
     "formik": "^2.2.9",
+    "js-cookie": "^3.0.5",
     "next": "13.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sharp": "^0.32.0",
+    "sweetalert2": "^11.7.5",
     "sweetalert2-react-content": "^5.0.7",
     "swiper": "^9.2.4",
     "typescript": "5.0.4"
