--- pages/application.tsx
@@ -1,8 +1,0 @@
-import ApplicationCard from '@/components/Application-Card';
-export default function Application() {
-  return (
-    <>
-      <ApplicationCard />
-    </>
-  );
-}
